Refactored benchmark-related code.
